requirements.txt
Django==5.0.8
django-ninja==1.3.0
psycopg[binary]==3.2.1
celery==5.4.0
redis==5.0.8
flower==2.0.1
python-dotenv==1.0.1
requests==2.32.3
chardet<6
pydantic==2.9.2
reportlab==4.2.2
weasyprint==62.3
pydyf==0.10.0
markdown==3.7
gunicorn==22.0.0
docxtpl==0.19.1
deep-translator==1.11.4
